pdf-chat-agent/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
passlib
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
langchain
langchain-community
langchain-huggingface
langchain-ollama
sentence-transformers
pdfminer.six
faiss-cpu
sse-starlette